refactor(infrastructure): reorganize storage adapters and providers
