+++ src/infrastructure/adapters/serialization/serialization.adapter.ts
