+++ src/infrastructure/adapters/web-storage.adapter.ts
